Set up and documented src/index.js to be imported
Fixes #8.
