+++ .eslintrc.js
@@ -14,7 +14,6 @@ module.exports = {
     rules: {
         "class-methods-use-this": 0,
         "import/extensions": 0,
-        "import/prefer-default-export": 0,
         "no-param-reassign": 0,
     },
 };
