+++ src/index.js
@@ -1,7 +1,30 @@
-import { launch40Phaser } from "./launch";
+import options from "./options.js";
+import EndScene from "./scenes/end.js";
+import GameScene from "./scenes/game.js";
+import StartScene from "./scenes/start.js";
 
-window.addEventListener("load", () => {
-    launch40Phaser({
-        debug: true,
+const defaultGameSettings = {
+    debug: false,
+    phaser: window.Phaser,
+};
+
+export const launch40Phaser = (setting = defaultGameSettings) => {
+    setting = { ...defaultGameSettings, ...setting };
+
+    return new setting.phaser.Game({
+        type: Phaser.AUTO,
+        width: options.windowWidth,
+        height: options.windowHeight,
+        backgroundColor: "#f9f9f9",
+        physics: {
+            default: "arcade",
+            arcade: {
+                debug: setting.debug,
+                gravity: { y: options.gravity },
+            },
+        },
+        scene: [StartScene, GameScene, EndScene],
     });
-});
+};
+
+export default launch40Phaser;
